Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -5,9 +5,9 @@ char* welcome() {
 }
 
 int main() {
-    int r;
-    scanf("%d",&r);
-    int ar=3.14*r*r;
-    printf("Area: %d",ar);
+    float r;
+    scanf("%f",&r);
+    float ar=3.14*r*r;
+    printf("Area: %.2f",ar);
     return 0;
 }
